Add files via upload
Hi...my name's Bhaswara this is a prototype Unity package for a Tycoon-like Game. Yes, i know, it heavily inspired from Indonesian Cuisine. that is the first idea i've had. apparently i've planned to released it on Web browser right now.

Future updates may be released, but i dunno if that's possible...it seems it can't put the entire project in it